requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
akshare>=1.14.0
tushare>=1.4.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
streamlit>=1.30.0
requests>=2.31.0
fastmcp>=0.1.0
python-docx>=1.0.0
playwright>=1.40.0
